fix: remove old CryptoV1
